手术记录仅录入者自己可见、添加医师组界面交互

+++ src/main/java/com/tongren/mapper/RecordMapper.java
@@ -21,6 +21,6 @@ public interface RecordMapper extends Mapper<Record> {
 
 	List<Record> selectDetailByFiltersForOthers(Map<String, Object> filters);
 
-	Integer selectTotalScore(Map<String, Object> filters);
+	Double selectTotalScore(Map<String, Object> filters);
 
 }
